Revert "Clear the send_window when we receive Fin packet"
This reverts commit 68c359783781edc3774247744b64cad2edc911de.

Indeed, the previous change would help to close the sockets sooner.

Indeed, the other end **might** not be nterested in the content of our
Data packets. But this is not true for the uTP specification. This is
true for the rust-utp design where we cannot issue concurrent actions.
Ideally we would be able to do something useful with the last data
received.