fix: tests
two of the tests were failing because they're badly designed. They're
designed under the assumption that no packet without a payload would
ever increase `seq_nr`. This is wrong because there is one packet
without payload that can cause `seq_nr` to be incremented.

At connection setup[1], a State packet, a packet with no payload is sent
by the server in response to a `Syn` packet sent by the client. The Syn
packet defines intent to establish a connection. The connection
establishment is aknowledged in the State packet sent back.

The first State packet that appears on an uTP connection is special
because it is aknowledged by the other end (i.e. its `seq_nr` value is
used to set `c.ack_nr`). Therefore, we need to increment the `seq_nr`
after sending the first State packet the connection ever sees.
Otherwise, one of the peers would make the wrong assumption that it has
already received a packet that it has never seen thanks to mismatched
`seq_nr` and `ack_nr`.

Another solution would be to set the "last acked" number to
(State.seq_nr - 1), but this solution (although cleaner and more
orthogonal to the rest of the specification) deviates from the
specification[1]. Therefore, this alternative solution is not used.

The commit that fixes the behaviour (but not the tests) is commit
53944ec86cb1af0c838f4dab5aa187a350b562cf.

Peter answered a few questions that helped to clarify the situation.

[1] http://www.bittorrent.org/beps/bep_0029.html#connection-setup