fix: `socket::test::test_receive_unexpected_reply_type_on_connect` test
Commit fd77155261c57e4c12aa991ed920d873b162ccaa introduced code that
made a test that was previously passing to fail. This usually means
regression.

I could very very easily add code to make the test pass, but the current
behaviour looks more right to me than the test described behaviour.
Therefore, I'm changing the test.

I won't revert the commit that introduced the "regression" because it
fixes an important bug that wasn't tested. Peter found a lot of issues
when he started emulating dropped packets and writing
more-than-just-unit-tests (something in the direction of integration
tests). Something that was seemingly being neglected previously.